requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
tqdm
pypdf
docx2txt
beautifulsoup4
azure-identity
azure-search-documents
python-dotenv
